requirements.txt
openai>=1.0.0
anthropic>=0.25.0
pyyaml>=6.0
python-dotenv>=1.0.0
weasyprint>=60.0
pyperclip>=1.8.0
click>=8.0